Improved infrastructure layer
