+++ backend/MusicInteraction/MusicInteraction.Infrastructure/MongoDB/MongoRepositories/Abstractions/IGradingMethodRepository.cs
@@ -1,0 +1,9 @@
+using MusicInteraction.Infrastructure.MongoDB.Entities;
+
+namespace MusicInteraction.Infrastructure.MongoDB;
+
+public interface IGradingMethodRepository : IMongoRepository<GradingMethodEntity>
+{
+    Task<List<GradingMethodEntity>> GetPublicGradingMethodsAsync();
+    Task<List<GradingMethodEntity>> GetUserGradingMethodsAsync(string userId);
+}
