+++ backend/MusicInteraction/MusicInteraction.Infrastructure/MongoDB/MongoRepositories/GradingMethodRepository.cs
@@ -3,12 +3,6 @@ using MusicInteraction.Infrastructure.MongoDB.Entities;
 
 namespace MusicInteraction.Infrastructure.MongoDB;
 
-public interface IGradingMethodRepository : IMongoRepository<GradingMethodEntity>
-{
-    Task<List<GradingMethodEntity>> GetPublicGradingMethodsAsync();
-    Task<List<GradingMethodEntity>> GetUserGradingMethodsAsync(string userId);
-}
-
 public class GradingMethodRepository : MongoRepository<GradingMethodEntity>, IGradingMethodRepository
 {
     private readonly IMongoCollection<GradingMethodEntity> _collection;
