+++ backend/MusicInteraction/MusicInteraction.Infrastructure/MongoDB/MongoRepositories/MongoDBRepository.cs
@@ -3,17 +3,6 @@ using MusicInteraction.Infrastructure.MongoDB.Entities;
 
 namespace MusicInteraction.Infrastructure.MongoDB;
 
-public interface IMongoRepository<TEntity> where TEntity : MongoEntity
-{
-    Task<List<TEntity>> GetAllAsync();
-    Task<TEntity> GetByIdAsync(Guid id);
-    Task<List<TEntity>> FindAsync(FilterDefinition<TEntity> filter);
-    Task CreateAsync(TEntity entity);
-    Task UpdateAsync(TEntity entity);
-    Task DeleteAsync(Guid id);
-    Task<bool> CollectionIsEmpty();
-}
-
 public class MongoRepository<TEntity> : IMongoRepository<TEntity> where TEntity : MongoEntity
 {
     private readonly IMongoCollection<TEntity> _collection;
